+++ milk-planning/src/test/scala/dev/atedeg/mdm/milkplanning/ActionsTest.scala
@@ -1,4 +1,4 @@
-package dev.atedeg.mdm.milkplanning.types
+package dev.atedeg.mdm.milkplanning
 
 import java.time.LocalDateTime
