+++ stocking/src/test/scala/dev/atedeg/mdm/stocking/Tests.scala
@@ -1,4 +1,4 @@
-package dev.atedeg.mdm.stocking.types
+package dev.atedeg.mdm.stocking
 
 import java.util.UUID
